tests/wpt/tests/css/css-mixins/mixin-cross-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Mixins: Mixin from another stylesheet</title>
    <link rel="help" href="https://drafts.csswg.org/css-mixins-1/#using-mixins">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      div {
        color: red;
        @apply --m1;
      }
    </style>
    <style>
      @mixin --m1() {
        color: green;
      }
    </style>
  </head>
  <body>
    <div><div class="cls" id="target">This text should be green.</div></div>
  <script>
    test(() => {
      let target = document.getElementById('target');
      assert_equals(getComputedStyle(target).color, 'rgb(0, 128, 0)');
    });
  </script>
  </body>
</html>
